patcher: move PSX S-commands to common
