+++ common/include/psxinit.h
@@ -1,0 +1,7 @@
+#ifndef _PSXINIT_H_
+#define _PSXINIT_H_
+
+// Reboots IOP with OSDSYS modules and switches PSX into PS2 mode.
+void switchPSX();
+
+#endif
