+++ patcher/CMakeLists.txt
@@ -61,6 +61,7 @@ set(EE_SOURCES
     ${PATCHER_SOURCE_DIR}/src/patches_osd.c
     ${PATCHER_SOURCE_DIR}/src/psx.c
     ${PATCHER_SOURCE_DIR}/../common/src/dprintf.c
+    ${PATCHER_SOURCE_DIR}/../common/src/psxinit.c
 )
 
 # HOSDMenu build
